Refactor Marketing screen to be available as a standalone fragment
It's unfortunately required.
